+++ source/base.html
@@ -83,7 +83,7 @@
 </footer>
 {% block pop_up %}
 {% endblock %}
-<script src='{% static "js/main.js" %}' type="module"></script>
+<script src='{% static "js/menu-toggle.js" %}'></script>
 </body>
 </html>
