+++ source/index.html
@@ -68,4 +68,5 @@
 {% block pop_up %}
   {% include 'popup-login.html' %}
   {% include 'popup-provost.html' %}
+  <script src="/static/js/main.js" type="module"></script>
 {% endblock %}
