--- source/js/login-popup.js
@@ -1,12 +1,0 @@
-const popup = document.querySelector('.popup-login');
-const login = document.getElementById('loginResult');
-popup.classList.add('visually-hidden');
-console.log(login.textContent);
-
-if (login.textContent === '1') {
-  popup.classList.remove('visually-hidden');
-  window.setTimeout(addHidden, 5000);
-  function addHidden() {
-    return popup.classList.add('visually-hidden');
-  }
-}
