+++ source/js/popup-login.js
@@ -1,0 +1,8 @@
+import {popup} from './popup.js'
+
+const popupLogin = document.querySelector('.popup-login');
+const login = document.getElementById('loginResult');
+
+if (login.textContent === '1') {
+  popup(popupLogin);
+}
